requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
mysql-connector-python==8.2.0

# Validation
pydantic==2.5.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2  # Required for TestClient in FastAPI

# CORS Support (included with FastAPI but listed for clarity)
# fastapi includes starlette which provides CORSMiddleware
